refactor: rename users to user
