+++ apps/gql/loaders/index.ts
@@ -1,11 +1,11 @@
 import { type Clients } from "~/clients";
 import { createSozlukLoaders } from "./sozluk";
 import { type DataLoaders } from "./types";
-import { createUsersLoader } from "./users";
+import { createUsersLoader } from "./user";
 
 export const createLoaders = (clients: Clients): DataLoaders => {
   return {
-    users: createUsersLoader(clients),
+    user: createUsersLoader(clients),
     sozluk: createSozlukLoaders(clients),
   };
 };
