+++ apps/gql/loaders/types.ts
@@ -1,8 +1,8 @@
 import { type SozlukTermLoader, type SozlukTermsLoader } from "./sozluk";
-import { type UsersLoader } from "./users";
+import { type UsersLoader } from "./user";
 
 export interface DataLoaders {
-  users: UsersLoader;
+  user: UsersLoader;
   sozluk: {
     term: SozlukTermLoader;
     terms: SozlukTermsLoader;
