+++ apps/gql/schema/resolvers/Query/user.ts
@@ -1,13 +1,13 @@
-import { UserLoaderKey } from "~/loaders/users";
+import { UserLoaderKey } from "~/loaders/user";
 import { type QueryResolvers } from "~/schema/types.generated";
 
 export const user: QueryResolvers["user"] = (_, args, { loaders }) => {
   if (args.id) {
-    return loaders.users.load(new UserLoaderKey("id", args.id));
+    return loaders.user.load(new UserLoaderKey("id", args.id));
   }
 
   if (args.username) {
-    return loaders.users.load(new UserLoaderKey("username", args.username));
+    return loaders.user.load(new UserLoaderKey("username", args.username));
   }
 
   throw new Error("id or username is required");
